requirements.txt
yfinance
matplotlib
pandas
requests
flask
mysql-connector-python
dotenv
apscheduler
pytz